+++ src/main/java/com/almightyalpaca/jetbrains/plugins/discord/settings/data/ApplicationSettings.java
@@ -49,4 +49,6 @@ public interface ApplicationSettings extends Settings
     boolean isShowFiles();
 
     boolean isShowElapsedTime();
+
+    boolean isForceBigIDEIcon();
 }